updater: Move automatic updates to its own package
