+++ internal/updater/automatic.go
@@ -1,4 +1,6 @@
-package main
+//go:build release
+
+package updater
 
 import (
 	"crypto/ed25519"
@@ -9,8 +11,8 @@ import (
 	"github.com/fynelabs/selfupdate"
 )
 
-// manageUpdates turns on automatic updates.
-func manageUpdates(a fyne.App, w fyne.Window) {
+// Enable turns on automatic application updates.
+func Enable(a fyne.App, w fyne.Window) {
 	publicKey := ed25519.PublicKey{165, 235, 49, 149, 238, 5, 192, 143, 38, 96, 124, 12, 235, 116, 94, 72, 244, 97, 230, 44, 60, 135, 85, 202, 98, 103, 233, 119, 63, 57, 83, 106}
 
 	// The public key above matches the signature of the below file served by our CDN
