+++ main.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/Jacalz/rymdport/v3/internal/assets"
 	"github.com/Jacalz/rymdport/v3/internal/ui"
+	"github.com/Jacalz/rymdport/v3/internal/updater"
 )
 
 func main() {
@@ -17,6 +18,6 @@ func main() {
 	w.Resize(fyne.NewSize(700, 400))
 	w.SetMaster()
 
-	manageUpdates(a, w)
+	updater.Enable(a, w)
 	w.ShowAndRun()
 }
